Refactor(workspace): Convered single package to multi-package workspace
