+++ Cargo.toml
@@ -1,53 +1,3 @@
-[package]
-name = "rusty-rolodex"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-async-trait = "0.1"
-chrono = { version = "0.4", features = ["serde"] }
-clap = {version = "4", features = ["derive", "env", "default"] }
-csv = "1.3"
-serde = {version = "1", features = ["derive"] }
-serde_json = "1"
-regex = "1"
-reqwest = {version = "0.13", features = ["blocking"]}
-dotenv = "0.15"
-rust-fuzzy-search = "0.1.1"
-tokio = { version = "1", features = ["full"] }
-toml = "0.8"
-url = "2.5"
-uuid = { version = "1", features = ["v4", "serde"] }
-
-[dev-dependencies]
-assert_cmd = "2.0"
-criterion = "0.5"
-mockito = "0.31"
-predicates = "3.1"
-tempfile = "3.6"
-rand = { version = "0.8", features = ["std_rng"] }
-
-
-[[bench]]
-name = "bench1k"
-harness = false
-
-[[bench]]
-name = "bench5k"
-harness = false
-
-[[bench]]
-name = "bench10k"
-harness = false
-
-[[bench]]
-name = "bench20k"
-harness = false
-
-[[bench]]
-name = "bench50k"
-harness = false
-
-[[bench]]
-name = "bench100k"
-harness = false
+[workspace]
+resolver = "3"
+members = ["cli","libs"]
