+++ cli/examples/add.sh
