+++ cli/examples/delete.sh
