+++ cli/examples/invalid_input.sh
