+++ cli/src/cli_component/mod.rs
@@ -1,4 +1,5 @@
 pub mod command;
 pub mod run;
 
+pub use libs::prelude::AppError;
 pub use run::run_app;
