+++ cli/src/cli_component/run.rs
@@ -1,13 +1,20 @@
-use crate::{
-    domain::contact,
-    prelude::{
-        AppError, ContactStore, RemoteStorage,
-        command::{Cli, Commands, ImportExportOption, SearchKey, SortKey},
-        contact::{Contact, EMAIL_REQ_MESSAGE, NAME_REQ_MESSAGE, PHONE_REQ_MESSAGE},
-        file::{CsvStorage, JsonStorage},
-        manager::{ContactManager, IndexUpdateType},
+use chrono::Utc;
+use crate::cli_component::command::{
+    Cli, Commands, ImportExportOption, 
+    SearchKey, SortKey
+};
+use libs::prelude::{
+    AppError,
+    ContactStore,
+    contact::{
+        Contact, EMAIL_REQ_MESSAGE,
+        NAME_REQ_MESSAGE, PHONE_REQ_MESSAGE,
     },
-    storage::{StorageMediums, remote::is_valid_url},
+    CsvStorage, JsonStorage,
+    manager::{ContactManager, IndexUpdateType},
+    RemoteStorage,
+    remote::is_valid_url,
+    StorageMediums,
 };
 use clap::Parser;
 use std::{env, path::Path, process::exit};
@@ -174,7 +181,7 @@ pub async fn run_app() -> Result<(), AppError> {
                     contact.tag = tag;
                 }
 
-                contact.updated_at = contact::Utc::now();
+                contact.updated_at = Utc::now();
             } else {
                 return Err(AppError::NotFound("Contact".to_string()));
             }
