+++ cli/tests/add.rs
