+++ cli/tests/delete.rs
