+++ cli/tests/edit_search.rs
