+++ cli/tests/list.rs
