+++ libs/src/domain/manager.rs
