+++ libs/src/domain/mod.rs
@@ -1,0 +1,7 @@
+pub mod contact;
+pub mod manager;
+
+use crate::errors::AppError;
+pub use contact::Contact;
+use crate::storage::{self, ContactStore, file};
+use uuid::Uuid;
