+++ libs/src/lib.rs
@@ -1,6 +1,5 @@
-pub mod cli;
 pub mod domain;
 pub mod errors;
 pub mod helper;
-pub mod prelude;
 pub mod storage;
+pub mod prelude;
