+++ libs/src/prelude.rs
@@ -1,0 +1,17 @@
+pub use crate::domain::{
+    contact::{self, Contact},
+    manager::{
+        self, ContactManager,
+    },
+};
+pub use crate::errors::AppError;
+pub use crate::storage::{
+    self, ContactStore, 
+    file::{
+        self, JsonStorage, TxtStorage, CsvStorage,
+    }, 
+    remote::{self, RemoteStorage,},
+    StorageMediums,
+};
+pub use std::collections::HashMap;
+pub use uuid;
