+++ libs/src/storage/memory.rs
