--- src/domain/mod.rs
@@ -1,4 +1,0 @@
-pub mod contact;
-pub mod manager;
-
-use crate::prelude::{AppError, Contact, ContactStore, file, storage, uuid::Uuid};
